fix pool candidates csv
